Refactor: modify API prefix and fix the bug of field type.

+++ sduoj-problem/src/main/java/cn/edu/sdu/qd/oj/problem/pojo/Problem.java
@@ -24,7 +24,7 @@ public class Problem implements Serializable {
 
     @Column(name = "p_is_public")
     @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-    private Byte isPublic;
+    private Integer isPublic;
 
     @Column(name = "u_id")
     private Integer userId;
